+++ packages/base/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.0]
+### Breaking
+- Migrated to `eslint` v9 and the new config format - `eslint.config.mjs` (https://eslint.org/docs/latest/use/configure/migration-guide)
+
 ## [1.1]
 ### Added
 - Bump `env` from `es2020` to `es2022`