+++ packages/typescript/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [3.0]
+### Breaking
+- Migrated to `eslint` v9 and new config format - `eslint.config.mjs` (https://eslint.org/docs/latest/use/configure/migration-guide)
+- Migrated to `typescript-eslint` v9 (https://github.com/typescript-eslint/typescript-eslint/releases/tag/v8.0.0)
+
 ## [2.0]
 ### Breaking
 - Bump `@typescript-eslint` from `v5` to `v6` and enable strict rules